requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
binance==0.3
binance-futures-connector==4.0.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dateparser==1.2.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.1.3
pandas==2.2.3
propcache==0.2.0
pycryptodome==3.21.0
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
tqdm==4.67.1
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.6
yarl==1.17.0
Flask[async]==2.3.3
